+++ apps/app/lib/app/features/home/widgets/home_featured_card.dart
