+++ apps/app/lib/app/features/list_items/cubit/list_items_cubit.dart
@@ -2,6 +2,9 @@ import 'package:bloc/bloc.dart';
 import 'package:core_models/core_models.dart';
 import 'package:listwhatever/app/features/list_items/cubit/list_items_state.dart';
 import 'package:listwhatever/app/features/list_items/repositories/list_items_repository.dart';
+import 'package:logging/logging.dart';
+
+final _log = Logger('ListItemsCubit');
 
 class ListItemsCubit extends Cubit<ListItemsState> {
   ListItemsCubit({required this._repository}) : super(const ListItemsState());
@@ -14,6 +17,7 @@ class ListItemsCubit extends Cubit<ListItemsState> {
       final items = await _repository.listByListId(listId);
       emit(ListItemsState(data: items));
     } catch (error) {
+      _log.severe('load error: $error');
       emit(ListItemsState(errorMessage: error.toString()));
     }
   }
@@ -21,14 +25,11 @@ class ListItemsCubit extends Cubit<ListItemsState> {
   Future<void> create(String listId, ListItem item) async {
     emit(ListItemsState(data: state.data, isLoading: true));
     try {
-      print('create: $item');
       await _repository.create(listId: listId, item: item);
       final items = await _repository.listByListId(listId);
-      print('create items: $items');
-      print('create items: ${items.length}');
       emit(ListItemsState(data: items));
     } catch (error) {
-      print('create error: $error');
+      _log.severe('create error: $error');
       emit(ListItemsState(data: state.data, errorMessage: error.toString()));
     }
   }
@@ -40,6 +41,7 @@ class ListItemsCubit extends Cubit<ListItemsState> {
       final items = await _repository.listByListId(listId);
       emit(ListItemsState(data: items));
     } catch (error) {
+      _log.severe('update error: $error');
       emit(ListItemsState(data: state.data, errorMessage: error.toString()));
     }
   }
@@ -51,6 +53,7 @@ class ListItemsCubit extends Cubit<ListItemsState> {
       final items = await _repository.listByListId(listId);
       emit(ListItemsState(data: items));
     } catch (error) {
+      _log.severe('delete error: $error');
       emit(ListItemsState(data: state.data, errorMessage: error.toString()));
     }
   }
