+++ apps/app/lib/app/features/list_items/pages/list_items_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:listwhatever/app/features/list_items/cubit/list_items_cubit.dart';
 import 'package:listwhatever/app/features/list_items/cubit/list_items_state.dart';
 import 'package:listwhatever/app/features/list_items/widgets/list_items_skeleton.dart';
-import 'package:listwhatever/app/features/list_items/list_items_view.dart';
-import 'package:listwhatever/app/features/list_items/list_map_view.dart';
+import 'package:listwhatever/app/features/list_items/pages/list_items_view.dart';
+import 'package:listwhatever/app/features/list_items/pages/list_map_view.dart';
 import 'package:listwhatever/shared/cubit_helpers/state_switcher.dart';
 
 class ListItemsPage extends StatelessWidget {
