+++ apps/app/lib/app/routing/routes.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:listwhatever/app/features/add_item/add_item_page.dart';
-import 'package:listwhatever/app/features/home/home_page.dart';
+import 'package:listwhatever/app/features/home/pages/home_page.dart';
 import 'package:listwhatever/app/features/list_items/cubit/list_items_cubit.dart';
-import 'package:listwhatever/app/features/list_items/list_items_page.dart';
+import 'package:listwhatever/app/features/list_items/pages/list_items_page.dart';
 import 'package:listwhatever/app/features/list_items/repositories/list_items_repository.dart';
 import 'package:listwhatever/app/features/login/login_page.dart';
 import 'package:listwhatever/app/features/new_list/new_list_page.dart';
