+++ apps/app/patrol_test/home_smoke_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:listwhatever/app/features/home/home_page.dart';
+import 'package:listwhatever/app/features/home/pages/home_page.dart';
 import 'package:listwhatever/app/main/app_theme.dart';
 import 'package:listwhatever/i18n/strings.g.dart';
 import 'package:patrol/patrol.dart';
