+++ apps/app/test/goldens/home_page_test.dart
@@ -3,7 +3,7 @@ library;
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:listwhatever/app/features/home/home_page.dart';
+import 'package:listwhatever/app/features/home/pages/home_page.dart';
 
 import 'support/golden_screenshot_presets.dart';
 import 'support/golden_test_helpers.dart';
